refactor: extract shared utilities and consolidate duplicated code
- shared/name-pools.js: single source of truth for name pools + toList()
- shared/watch-utils.js: shared file watcher (replaces duplicated fs.watch loops)
- shared/mime-types.js: shared MIME type map (replaces inline maps in both servers)
- collector/index.ts: use shared/watch-utils.js, remove inline fs.watch loop
- hubreceiver/server.ts: use shared/mime-types.js
- claudeville/server.ts: use shared/watch-utils.js + shared/mime-types.js
- runtime-config.shared.ts: import from shared/name-pools.js
- claudeville/src/config/agentNames.ts: import from shared/name-pools.js
- test files: import from shared/cost.js instead of local copies
- vitest.config.ts: add resolve alias + create claudeville/shared symlink
