+++ claudeville/src/config/agentNames.ts
@@ -1,33 +1,11 @@
+import { DEFAULT_AGENT_NAME_POOL, DEFAULT_SESSION_NAME_POOL, toList } from '../../shared/name-pools.js';
 import { Appearance } from '../domain/value-objects/Appearance.js';
 import { getRuntimeConfig } from './runtime.js';
 
 const DEFAULT_AGENT_NAME_POOLS = {
-    agent: [
-        'Atlas', 'Nova', 'Cipher', 'Pixel', 'Spark',
-        'Bolt', 'Echo', 'Flux', 'Helix', 'Onyx',
-        'Prism', 'Qubit', 'Rune', 'Sage', 'Vex',
-    ],
-    session: [
-        'Orbit', 'Beacon', 'Relay', 'Pulse', 'Signal',
-        'Vector', 'Comet', 'Drift', 'Trace', 'Kernel',
-        'Node', 'Echo', 'Wisp', 'Shard', 'Tide',
-    ],
+  agent: DEFAULT_AGENT_NAME_POOL,
+  session: DEFAULT_SESSION_NAME_POOL,
 };
-
-function toList(value) {
-    if (Array.isArray(value)) {
-        return value.map((item) => String(item).trim()).filter(Boolean);
-    }
-
-    if (typeof value !== 'string') {
-        return [];
-    }
-
-    return value
-        .split(',')
-        .map((item) => item.trim())
-        .filter(Boolean);
-}
 
 function normalizeNamePools(rawPools: Record<string, unknown> = {}) {
     const agent = toList(rawPools.agent || rawPools.en || rawPools.agentNames || rawPools.agentNamePool);
